+++ bookmark.go
@@ -30,5 +30,6 @@ type Bookmark struct {
 	Module   string   `json:"module"`
 	Version  uint64   `json:"version"`
 	Modified uint64   `json:"modified"`
+	Xhsum    string   `json:"xhsum"`
 	//flags int
 }