+++ internal/api/api.go
@@ -20,15 +20,16 @@
 package api
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"net/http"
 	"strconv"
+	"strings"
+
+	"github.com/go-chi/chi/v5"
 
 	"github.com/blob42/gosuki"
 	db "github.com/blob42/gosuki/internal/database"
-	"github.com/go-chi/chi/v5"
 )
 
 type Bookmark = gosuki.Bookmark
@@ -41,37 +42,7 @@ type Payload struct {
 	Result  any  `json:"result"`
 }
 
-type ReqIsFuzzy struct{}
 type ResetPage struct{}
-
-func IsFuzzy(r *http.Request) bool {
-	fuzzy := r.Context().Value(ReqIsFuzzy{})
-
-	if v, ok := fuzzy.(bool); ok && v {
-		return true
-	}
-
-	return false
-}
-
-// Find and add fuzzy search parameter to the request context
-func trackFuzzySearch(r *http.Request) *http.Request {
-	var fuzzy bool
-
-	query := r.URL.Query().Get("query")
-
-	if fuzzyParam := r.URL.Query().Get("fuzzy"); fuzzyParam != "" {
-		fuzzy = true
-	}
-
-	// Check if the first character of query is `~`
-	if len(query) > 0 && query[0] == '~' {
-		fuzzy = true
-	}
-
-	rCtx := context.WithValue(r.Context(), ReqIsFuzzy{}, fuzzy)
-	return r.WithContext(rCtx)
-}
 
 func GetPaginationParams(r *http.Request) *db.PaginationParams {
 	pageParams := db.DefaultPagination()
@@ -134,13 +105,49 @@ func GetBookmarks(r *http.Request) ([]*gosuki.Bookmark, uint, error) {
 	}
 
 	pageParams := GetPaginationParams(r)
+	var (
+		searchByQuery = query != ""
+		searchByTag   = tag != ""
+	)
 
-	if query != "" && tag != "" {
-		qResult, err = db.QueryBookmarksByTag(r.Context(), query, tag, IsFuzzy(r), pageParams)
-	} else if tag != "" {
-		qResult, err = db.BookmarksByTag(r.Context(), tag, pageParams)
-	} else if query != "" {
+	// Search with query AND tags
+	if searchByQuery && searchByTag {
+		if strings.Contains(tag, ",") {
+			// Query with multiple tags
+			qResult, err = db.QueryBookmarksByTags(
+				r.Context(),
+				query,
+				strings.Split(tag, ","),
+				db.TagAnd, // AND tags
+				IsFuzzy(r),
+				pageParams,
+			)
+		} else {
+			qResult, err = db.QueryBookmarksByTag(r.Context(),
+				query,
+				tag,
+				IsFuzzy(r),
+				pageParams,
+			)
+		}
+
+		// Tag based search mode
+	} else if searchByTag {
+		//query with many tags
+		if strings.Contains(tag, ",") {
+			qResult, err = db.BookmarksByTags(
+				r.Context(),
+				strings.Split(tag, ","),
+				db.TagAnd,
+				pageParams,
+			)
+		} else {
+			qResult, err = db.BookmarksByTag(r.Context(), tag, pageParams)
+		}
+	} else if searchByQuery {
 		qResult, err = db.QueryBookmarks(r.Context(), query, IsFuzzy(r), pageParams)
+
+		// No filter mode, list all bookmarks paginated
 	} else {
 		qResult, err = db.ListBookmarks(r.Context(), pageParams)
 	}